kafka_streaming/requirements.txt
kafka-python-ng==2.2.3
cassandra-driver[gevent]==3.29.0
gevent>=23.9.1
faker==20.1.0
python-dotenv==1.0.0
six>=1.16.0